requirements.txt
pytorch-lightning>=1.3.8,<2
torch>=1.4
hydra-core>=1.1.0
fiftyone>=0.13
icevision>=0.8
lightning-flash[image]>=0.5.2